+++ src/main/java/org/markdownwriterfx/options/Options.java
@@ -59,6 +59,8 @@ public class Options
 	public static final int MAX_FONT_SIZE = 36;
 	public static final String DEF_MARKDOWN_FILE_EXTENSIONS = "*.md,*.markdown,*.txt";
 	public enum RendererType { CommonMark, FlexMark };
+	public static final int DEF_WRAP_LINE_LENGTH = 80;
+	public static final int MIN_WRAP_LINE_LENGTH = 10;
 
 	public static void load(Preferences options) {
 		fontFamily.init(options, "fontFamily", null, value -> safeFontFamily(value));
@@ -74,6 +76,8 @@ public class Options
 		emphasisMarker.init(options, "emphasisMarker", "_");
 		strongEmphasisMarker.init(options, "strongEmphasisMarker", "**");
 		bulletListMarker.init(options, "bulletListMarker", "-");
+
+		wrapLineLength.init(options, "wrapLineLength", DEF_WRAP_LINE_LENGTH);
 	}
 
 	/**
@@ -163,4 +167,10 @@ public class Options
 	public static String getBulletListMarker() { return bulletListMarker.get(); }
 	public static void setBulletListMarker(String bulletListMarker) { Options.bulletListMarker.set(bulletListMarker); }
 	public static StringProperty bulletListMarkerProperty() { return bulletListMarker; }
+
+	// 'wrapLineLength' property
+	private static final PrefsIntegerProperty wrapLineLength = new PrefsIntegerProperty();
+	public static int getWrapLineLength() { return wrapLineLength.get(); }
+	public static void setWrapLineLength(int wrapLineLength) { Options.wrapLineLength.set(Math.max(wrapLineLength, MIN_WRAP_LINE_LENGTH)); }
+	public static IntegerProperty wrapLineLengthProperty() { return wrapLineLength; }
 }